Remove conditional update logic
Comparing the file timestamps with the update times returned from the API
is not enough: In the case that you change the config file to use an
"older" secrets file to populate a secret that has been updated recently,
this would not work.

What would be needed is a change-time tracking for individual config
entries, which in turn means that you could not really manually update
the config but we'd need to support this with tooling.

So for now, just unconditionally update all the secrets and let's seen
at which point we hit rate limits.